Refactor Git handlers into modular domain files; consolidate GraphBucket
Split 2093-line git.ts into six focused modules (read, branch, streaming, worktree, worktreeWatch, github) with a thin composer in git/index.ts for clarity and maintainability. Consolidate GraphBucket type definition to shared/history.ts canonical location, replacing duplicates across director-notes, preload, and renderer components. Update IPC-PATTERNS.md to document the new organization.
