--- src/main/ipc/handlers/git.ts
@@ -1,2093 +1,0 @@
-import { ipcMain, type BrowserWindow } from 'electron';
-import fs from 'fs/promises';
-import path from 'path';
-import chokidar, { FSWatcher } from 'chokidar';
-import { execFileNoThrow, execFileBufferNoThrow, execFileStreaming } from '../../utils/execFile';
-import { execGit } from '../../utils/remote-git';
-import { buildSshCommand } from '../../utils/ssh-command-builder';
-import { logger } from '../../utils/logger';
-import { getSshRemoteById } from '../../stores';
-import { createSafeSend } from '../../utils/safe-send';
-import {
-	withIpcErrorLogging,
-	createIpcHandler,
-	CreateHandlerOptions,
-} from '../../utils/ipcHandler';
-import { resolveGhPath, getCachedGhStatus, setCachedGhStatus } from '../../utils/cliDetection';
-import { getShellPath } from '../../runtime/getShellPath';
-import { captureMessage } from '../../utils/sentry';
-import { WINDOWS_LOCKED_SYSTEM_FILES } from '../../utils/watcher-ignore';
-import {
-	parseGitBranches,
-	parseGitTags,
-	parseGitBehindAhead,
-	countUncommittedChanges,
-	isImageFile,
-	getImageMimeType,
-	isWorktreeAlreadyUsedError,
-	parseWorktreePathForBranch,
-} from '../../../shared/gitUtils';
-import type { GitRunCommandResult, GitStreamingOperation } from '../../../shared/gitUtils';
-import {
-	worktreeInfoRemote,
-	worktreeSetupRemote,
-	worktreeCheckoutRemote,
-	listWorktreesRemote,
-	getRepoRootRemote,
-} from '../../utils/remote-git';
-import { readDirRemote } from '../../utils/remote-fs';
-import { captureException } from '../../utils/sentry';
-import { markStaleForDeletedWorktreeUsingStore } from '../../agent-run/worktree-stale';
-
-const LOG_CONTEXT = '[Git]';
-
-/**
- * Directory-scan failures that are environmental rather than bugs: the path was
- * moved or deleted out from under us (ENOENT/ENOTDIR), or we don't hold read
- * permission on it (EACCES, or EPERM for macOS TCC-protected locations like
- * Documents and Desktop). The scan skips the directory and reports `scanFailed`;
- * none of these are worth a Sentry report.
- */
-const EXPECTED_SCAN_ERROR_CODES = new Set(['ENOENT', 'ENOTDIR', 'EACCES', 'EPERM']);
-
-function isExpectedScanError(code: string | undefined): boolean {
-	return code !== undefined && EXPECTED_SCAN_ERROR_CODES.has(code);
-}
-
-/**
- * Dependencies for Git handlers
- */
-export interface GitHandlerDependencies {
-	/** Settings store for accessing SSH remote configurations */
-	settingsStore: {
-		get: (key: string, defaultValue?: unknown) => unknown;
-	};
-	/**
-	 * Returns the current main window (or null). Used to route worktree
-	 * watcher events through safeSend so web-desktop bridge clients receive
-	 * them alongside the desktop renderer.
-	 */
-	getMainWindow: () => BrowserWindow | null;
-}
-
-// Worktree directory watchers keyed by session ID
-const worktreeWatchers = new Map<string, FSWatcher>();
-// Debounce timers keyed by "sessionId:dirPath" so each discovered directory
-// gets its own independent timer (previously keyed by sessionId alone, which
-// caused only the last of multiple near-simultaneous addDir events to fire).
-const worktreeWatchDebounceTimers = new Map<string, NodeJS.Timeout>();
-
-/** Helper to create handler options with Git context */
-const handlerOpts = (operation: string, logSuccess = false): CreateHandlerOptions => ({
-	context: LOG_CONTEXT,
-	operation,
-	logSuccess,
-});
-
-/**
- * Look up the worktree path currently checked out on the given branch
- * by running `git worktree list --porcelain` against the local repo.
- *
- * Used to recover from `git worktree add` failures with the "already used /
- * already checked out" error: instead of bubbling up an opaque error, we
- * return the existing worktree path so callers can open it as a session.
- *
- * Stale registrations (where the directory was deleted manually without
- * `git worktree prune`) are filtered out by an `fs.access` check so callers
- * never get a path that points at nothing.
- *
- * @returns Absolute worktree path, or null if not found / stale
- */
-async function findLocalWorktreeForBranch(
-	mainRepoCwd: string,
-	branchName: string
-): Promise<string | null> {
-	const result = await execFileNoThrow('git', ['worktree', 'list', '--porcelain'], mainRepoCwd);
-	if (result.exitCode !== 0) return null;
-	const existingPath = parseWorktreePathForBranch(result.stdout, branchName);
-	if (!existingPath) return null;
-	try {
-		await fs.access(existingPath);
-		return existingPath;
-	} catch {
-		return null;
-	}
-}
-
-// Cancel callbacks for in-flight streaming git commands, keyed by runId.
-const streamingGitRuns = new Map<string, () => void>();
-
-/**
- * Build the argv for a streaming git operation.
- *
- * `--progress` is required because git suppresses transfer progress when stdout
- * isn't a TTY, which is exactly our case - without it the modal would sit blank
- * until the command finished.
- */
-function buildStreamingGitArgs(
-	operation: GitStreamingOperation,
-	branch: string | null,
-	setUpstream: boolean
-): string[] {
-	if (operation === 'push') {
-		return setUpstream && branch
-			? ['push', '--progress', '--set-upstream', 'origin', branch]
-			: ['push', '--progress'];
-	}
-	return [operation, '--progress'];
-}
-
-/**
- * Run a network git operation, forwarding output to the requesting window as it
- * arrives and resolving once the process exits.
- */
-async function runStreamingGitCommand(
-	event: Electron.IpcMainInvokeEvent,
-	options: {
-		runId: string;
-		operation: GitStreamingOperation;
-		cwd: string;
-		sshRemoteId?: string;
-		remoteCwd?: string;
-		setUpstream?: boolean;
-	}
-): Promise<GitRunCommandResult> {
-	const { runId, operation, cwd, sshRemoteId, remoteCwd, setUpstream = false } = options;
-
-	const sshRemote = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-	// The user explicitly opted into SSH - running against the local repo instead
-	// would push/pull the wrong tree, so fail loudly.
-	if (sshRemoteId && !sshRemote) {
-		return {
-			success: false,
-			exitCode: 1,
-			cancelled: false,
-			error: `SSH remote not found: ${sshRemoteId}`,
-		};
-	}
-	const effectiveRemoteCwd = sshRemote ? remoteCwd || cwd : undefined;
-
-	// `--set-upstream` needs the branch name spelled out.
-	let branch: string | null = null;
-	if (operation === 'push' && setUpstream) {
-		const branchResult = await execGit(
-			['rev-parse', '--abbrev-ref', 'HEAD'],
-			cwd,
-			sshRemote,
-			effectiveRemoteCwd
-		);
-		branch = branchResult.exitCode === 0 ? branchResult.stdout.trim() : null;
-		if (!branch) {
-			return {
-				success: false,
-				exitCode: 1,
-				cancelled: false,
-				error: 'Could not determine the current branch to set upstream for',
-			};
-		}
-	}
-
-	const gitArgs = buildStreamingGitArgs(operation, branch, setUpstream);
-
-	// Full shell PATH so git hooks (Husky pre-push running npm, etc.) resolve
-	// their tooling, and GIT_TERMINAL_PROMPT=0 so a missing credential fails
-	// fast with a readable error instead of hanging on a prompt nobody can see.
-	let env: NodeJS.ProcessEnv = { ...process.env, GIT_TERMINAL_PROMPT: '0' };
-	try {
-		const shellPath = await getShellPath();
-		if (shellPath) env = { ...env, PATH: shellPath };
-	} catch (error) {
-		captureMessage(
-			`git:runCommand falling back to default PATH: ${error instanceof Error ? error.message : String(error)}`,
-			'warning'
-		);
-	}
-
-	let command = 'git';
-	let args = gitArgs;
-	let spawnCwd: string | undefined = cwd;
-	if (sshRemote) {
-		const sshCommand = await buildSshCommand(sshRemote, {
-			command: 'git',
-			args: gitArgs,
-			cwd: effectiveRemoteCwd,
-			env: sshRemote.remoteEnv,
-		});
-		command = sshCommand.command;
-		args = sshCommand.args;
-		// The remote cwd is applied inside the SSH payload; the local process
-		// shouldn't inherit a path that may not exist on this machine.
-		spawnCwd = undefined;
-	}
-
-	const send = (chunk: string, stream: 'stdout' | 'stderr') => {
-		if (event.sender.isDestroyed()) return;
-		event.sender.send('git:commandOutput', { runId, stream, chunk });
-	};
-
-	const handle = execFileStreaming(command, args, {
-		cwd: spawnCwd,
-		env,
-		onChunk: send,
-	});
-	streamingGitRuns.set(runId, handle.cancel);
-
-	try {
-		const result = await handle.result;
-		const cancelled = result.exitCode === 'SIGTERM';
-		return {
-			success: result.exitCode === 0,
-			exitCode: result.exitCode,
-			cancelled,
-			error:
-				result.exitCode === 0 || cancelled
-					? undefined
-					: result.stderr.trim() || `git ${operation} exited with ${result.exitCode}`,
-		};
-	} finally {
-		streamingGitRuns.delete(runId);
-	}
-}
-
-/**
- * Register all Git-related IPC handlers.
- *
- * These handlers provide Git operations used across the application including:
- * - Basic operations: status, diff, branch, remote, tags
- * - Advanced queries: log, info, commitCount
- * - File operations: show, showFile
- * - Worktree management: worktreeInfo, worktreeSetup, worktreeCheckout (with SSH support)
- * - GitHub CLI integration: checkGhCli, createPR, getDefaultBranch
- *
- * @param deps Dependencies including settingsStore for SSH remote configuration lookup
- */
-export function registerGitHandlers(deps: GitHandlerDependencies): void {
-	const safeSend = createSafeSend(deps.getMainWindow);
-
-	// Basic Git operations
-	// All handlers accept optional sshRemoteId and remoteCwd for remote execution
-
-	// --- FIX: Always pass cwd as remoteCwd for remote git operations ---
-	ipcMain.handle(
-		'git:status',
-		withIpcErrorLogging(
-			handlerOpts('status'),
-			async (cwd: string, sshRemoteId?: string, remoteCwd?: string) => {
-				const sshRemote = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-				const effectiveRemoteCwd = sshRemote ? remoteCwd || cwd : undefined;
-				const result = await execGit(['status', '--porcelain'], cwd, sshRemote, effectiveRemoteCwd);
-				return { stdout: result.stdout, stderr: result.stderr };
-			}
-		)
-	);
-
-	ipcMain.handle(
-		'git:diff',
-		withIpcErrorLogging(
-			handlerOpts('diff'),
-			async (cwd: string, file?: string, sshRemoteId?: string, remoteCwd?: string) => {
-				const args = file ? ['diff', file] : ['diff'];
-				const sshRemote = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-				const effectiveRemoteCwd = sshRemote ? remoteCwd || cwd : undefined;
-				const result = await execGit(args, cwd, sshRemote, effectiveRemoteCwd);
-				return { stdout: result.stdout, stderr: result.stderr };
-			}
-		)
-	);
-
-	ipcMain.handle(
-		'git:isRepo',
-		withIpcErrorLogging(
-			handlerOpts('isRepo'),
-			async (cwd: string, sshRemoteId?: string, remoteCwd?: string) => {
-				const sshRemote = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-				const effectiveRemoteCwd = sshRemote ? remoteCwd || cwd : undefined;
-				const result = await execGit(
-					['rev-parse', '--is-inside-work-tree'],
-					cwd,
-					sshRemote,
-					effectiveRemoteCwd
-				);
-				return result.exitCode === 0;
-			}
-		)
-	);
-
-	// Initialize a new git repository at the given directory.
-	// Returns { success, error? }. Used by the agent-create UI to offer
-	// `git init` when a chosen working directory isn't already a repo.
-	ipcMain.handle(
-		'git:init',
-		withIpcErrorLogging(
-			handlerOpts('init'),
-			async (cwd: string, sshRemoteId?: string, remoteCwd?: string) => {
-				const sshRemote = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-				// Fail fast if an SSH remote was requested but can't be resolved -
-				// otherwise we'd silently `git init` the wrong (local) directory.
-				if (sshRemoteId && !sshRemote) {
-					return {
-						success: false,
-						error: `SSH remote not found: ${sshRemoteId}`,
-					};
-				}
-				const effectiveRemoteCwd = sshRemote ? remoteCwd || cwd : undefined;
-				const result = await execGit(['init'], cwd, sshRemote, effectiveRemoteCwd);
-				if (result.exitCode !== 0) {
-					return {
-						success: false,
-						error: result.stderr?.trim() || 'git init failed',
-					};
-				}
-				return { success: true };
-			}
-		)
-	);
-
-	// Stage every change (new, modified, deleted) and commit it in one shot.
-	// Used by Auto Run to checkpoint each iteration. Returns
-	// { success, committed, commitHash?, error? }. A clean tree is NOT an error:
-	// it resolves { success: true, committed: false } so callers can no-op
-	// quietly. Commit failures (e.g. missing git identity, failing hooks) surface
-	// as { success: false } so the caller can log without aborting its run.
-	ipcMain.handle(
-		'git:commitAll',
-		withIpcErrorLogging(
-			handlerOpts('commitAll'),
-			async (cwd: string, message: string, sshRemoteId?: string, remoteCwd?: string) => {
-				const sshRemote = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-				// Fail fast if an SSH remote was requested but can't be resolved -
-				// otherwise we'd silently commit in the wrong (local) directory.
-				if (sshRemoteId && !sshRemote) {
-					return {
-						success: false,
-						committed: false,
-						error: `SSH remote not found: ${sshRemoteId}`,
-					};
-				}
-				const effectiveRemoteCwd = sshRemote ? remoteCwd || cwd : undefined;
-
-				// Stage everything first so the porcelain check below reflects the
-				// full pending change set (including untracked files).
-				const addResult = await execGit(['add', '-A'], cwd, sshRemote, effectiveRemoteCwd);
-				if (addResult.exitCode !== 0) {
-					return {
-						success: false,
-						committed: false,
-						error: addResult.stderr?.trim() || 'git add failed',
-					};
-				}
-
-				// Nothing staged → clean tree, nothing to commit. Report success.
-				const statusResult = await execGit(
-					['status', '--porcelain'],
-					cwd,
-					sshRemote,
-					effectiveRemoteCwd
-				);
-				if (statusResult.exitCode === 0 && statusResult.stdout.trim() === '') {
-					return { success: true, committed: false };
-				}
-
-				const commitResult = await execGit(
-					['commit', '-m', message],
-					cwd,
-					sshRemote,
-					effectiveRemoteCwd
-				);
-				if (commitResult.exitCode !== 0) {
-					return {
-						success: false,
-						committed: false,
-						error:
-							commitResult.stderr?.trim() || commitResult.stdout?.trim() || 'git commit failed',
-					};
-				}
-
-				// Best-effort short hash for feedback/logging; absence is non-fatal.
-				const hashResult = await execGit(
-					['rev-parse', '--short', 'HEAD'],
-					cwd,
-					sshRemote,
-					effectiveRemoteCwd
-				);
-				const commitHash = hashResult.exitCode === 0 ? hashResult.stdout.trim() : undefined;
-				return { success: true, committed: true, commitHash };
-			}
-		)
-	);
-
-	ipcMain.handle(
-		'git:numstat',
-		withIpcErrorLogging(
-			handlerOpts('numstat'),
-			async (cwd: string, sshRemoteId?: string, remoteCwd?: string) => {
-				const sshRemote = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-				const effectiveRemoteCwd = sshRemote ? remoteCwd || cwd : undefined;
-				const result = await execGit(['diff', '--numstat'], cwd, sshRemote, effectiveRemoteCwd);
-				return { stdout: result.stdout, stderr: result.stderr };
-			}
-		)
-	);
-
-	ipcMain.handle(
-		'git:branch',
-		withIpcErrorLogging(
-			handlerOpts('branch'),
-			async (cwd: string, sshRemoteId?: string, remoteCwd?: string) => {
-				const sshRemote = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-				const effectiveRemoteCwd = sshRemote ? remoteCwd || cwd : undefined;
-				const result = await execGit(
-					['rev-parse', '--abbrev-ref', 'HEAD'],
-					cwd,
-					sshRemote,
-					effectiveRemoteCwd
-				);
-				return { stdout: result.stdout.trim(), stderr: result.stderr };
-			}
-		)
-	);
-
-	ipcMain.handle(
-		'git:remote',
-		withIpcErrorLogging(
-			handlerOpts('remote'),
-			async (cwd: string, sshRemoteId?: string, remoteCwd?: string) => {
-				const sshRemote = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-				const effectiveRemoteCwd = sshRemote ? remoteCwd || cwd : undefined;
-				const result = await execGit(
-					['remote', 'get-url', 'origin'],
-					cwd,
-					sshRemote,
-					effectiveRemoteCwd
-				);
-				return { stdout: result.stdout.trim(), stderr: result.stderr };
-			}
-		)
-	);
-
-	ipcMain.handle(
-		'git:branches',
-		withIpcErrorLogging(
-			handlerOpts('branches'),
-			async (cwd: string, sshRemoteId?: string, remoteCwd?: string) => {
-				const sshRemote = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-				const effectiveRemoteCwd = sshRemote ? remoteCwd || cwd : undefined;
-				const result = await execGit(
-					['branch', '-a', '--format=%(refname:short)'],
-					cwd,
-					sshRemote,
-					effectiveRemoteCwd
-				);
-				if (result.exitCode !== 0) {
-					return { branches: [], stderr: result.stderr };
-				}
-				// Use shared parsing function
-				const branches = parseGitBranches(result.stdout);
-				return { branches };
-			}
-		)
-	);
-
-	ipcMain.handle(
-		'git:tags',
-		withIpcErrorLogging(
-			handlerOpts('tags'),
-			async (cwd: string, sshRemoteId?: string, remoteCwd?: string) => {
-				const sshRemote = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-				const effectiveRemoteCwd = sshRemote ? remoteCwd || cwd : undefined;
-				const result = await execGit(['tag', '--list'], cwd, sshRemote, effectiveRemoteCwd);
-				if (result.exitCode !== 0) {
-					return { tags: [], stderr: result.stderr };
-				}
-				// Use shared parsing function
-				const tags = parseGitTags(result.stdout);
-				return { tags };
-			}
-		)
-	);
-
-	ipcMain.handle(
-		'git:info',
-		withIpcErrorLogging(
-			handlerOpts('info'),
-			async (cwd: string, sshRemoteId?: string, remoteCwd?: string) => {
-				const sshRemote = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-				const effectiveRemoteCwd = sshRemote ? remoteCwd || cwd : undefined;
-				// Get comprehensive git info in a single call
-				const [branchResult, remoteResult, statusResult, behindAheadResult] = await Promise.all([
-					execGit(['rev-parse', '--abbrev-ref', 'HEAD'], cwd, sshRemote, effectiveRemoteCwd),
-					execGit(['remote', 'get-url', 'origin'], cwd, sshRemote, effectiveRemoteCwd),
-					execGit(['status', '--porcelain'], cwd, sshRemote, effectiveRemoteCwd),
-					execGit(
-						['rev-list', '--left-right', '--count', '@{upstream}...HEAD'],
-						cwd,
-						sshRemote,
-						effectiveRemoteCwd
-					),
-				]);
-
-				// Use shared parsing functions for behind/ahead and uncommitted changes
-				const { behind, ahead } =
-					behindAheadResult.exitCode === 0
-						? parseGitBehindAhead(behindAheadResult.stdout)
-						: { behind: 0, ahead: 0 };
-				const uncommittedChanges = countUncommittedChanges(statusResult.stdout);
-
-				return {
-					branch: branchResult.stdout.trim(),
-					remote: remoteResult.stdout.trim(),
-					behind,
-					ahead,
-					uncommittedChanges,
-				};
-			}
-		)
-	);
-
-	ipcMain.handle(
-		'git:log',
-		withIpcErrorLogging(
-			handlerOpts('log'),
-			async (
-				cwd: string,
-				options?: { limit?: number; search?: string },
-				sshRemoteId?: string,
-				remoteCwd?: string
-			) => {
-				const sshRemote = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-				const effectiveRemoteCwd = sshRemote ? remoteCwd || cwd : undefined;
-				// Get git log with formatted output for parsing
-				// Format: hash|author|date|refs|subject followed by shortstat
-				// Using a unique separator to split commits
-				const limit = options?.limit || 100;
-				const args = [
-					'log',
-					`--max-count=${limit}`,
-					'--pretty=format:COMMIT_START%H|%an|%ad|%D|%s',
-					'--date=iso-strict',
-					'--shortstat',
-				];
-
-				// Add search filter if provided
-				if (options?.search) {
-					args.push('--all', `--grep=${options.search}`, '-i');
-				}
-
-				const result = await execGit(args, cwd, sshRemote, effectiveRemoteCwd);
-
-				if (result.exitCode !== 0) {
-					return { entries: [], error: result.stderr };
-				}
-
-				// Split by COMMIT_START marker and parse each commit
-				const commits = result.stdout.split('COMMIT_START').filter((c) => c.trim());
-				const entries = commits.map((commitBlock) => {
-					const lines = commitBlock.split('\n').filter((l) => l.trim());
-					const mainLine = lines[0];
-					const [hash, author, date, refs, ...subjectParts] = mainLine.split('|');
-
-					// Parse shortstat line (e.g., " 3 files changed, 10 insertions(+), 5 deletions(-)")
-					let additions = 0;
-					let deletions = 0;
-					const statLine = lines.find((l) => l.includes('changed'));
-					if (statLine) {
-						const addMatch = statLine.match(/(\d+) insertion/);
-						const delMatch = statLine.match(/(\d+) deletion/);
-						if (addMatch) additions = parseInt(addMatch[1], 10);
-						if (delMatch) deletions = parseInt(delMatch[1], 10);
-					}
-
-					return {
-						hash,
-						shortHash: hash?.slice(0, 7),
-						author,
-						date,
-						refs: refs ? refs.split(', ').filter((r) => r.trim()) : [],
-						subject: subjectParts.join('|'), // In case subject contains |
-						additions,
-						deletions,
-					};
-				});
-
-				return { entries, error: null };
-			}
-		)
-	);
-
-	// Topology data for graph view: includes parent hashes for lane rendering.
-	ipcMain.handle(
-		'git:graph',
-		withIpcErrorLogging(
-			handlerOpts('graph'),
-			async (
-				cwd: string,
-				options?: { limit?: number },
-				sshRemoteId?: string,
-				remoteCwd?: string
-			) => {
-				const sshRemote = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-				const effectiveRemoteCwd = sshRemote ? remoteCwd || cwd : undefined;
-				const limit = options?.limit || 200;
-				// Use ASCII Unit Separator (U+001F, written as %x1f in git's pretty-format)
-				// between fields. `|` was tempting but author names and subjects can legally
-				// contain it, which silently corrupts every field after the offending one.
-				// US is a non-printing control character that never appears in real text.
-				const args = [
-					'log',
-					'--all',
-					`--max-count=${limit}`,
-					'--pretty=format:GRAPH_START%H%x1f%P%x1f%an%x1f%ad%x1f%D%x1f%s',
-					'--date=iso-strict',
-				];
-				const result = await execGit(args, cwd, sshRemote, effectiveRemoteCwd);
-				if (result.exitCode !== 0) {
-					return { nodes: [], error: result.stderr };
-				}
-				const nodes = result.stdout
-					.split('GRAPH_START')
-					.filter((c) => c.trim())
-					.map((block) => {
-						const trimmed = block.trim();
-						const [hash = '', parents = '', author = '', date = '', refs = '', ...subj] =
-							trimmed.split('\x1f');
-						return {
-							hash,
-							shortHash: hash.slice(0, 7),
-							parents: parents ? parents.split(' ').filter(Boolean) : [],
-							author,
-							date,
-							refs: refs ? refs.split(', ').filter((r) => r.trim()) : [],
-							// Re-join with the same separator in case the subject itself contained one
-							// (extremely unlikely for a control character, but cheap to be correct).
-							subject: subj.join('\x1f'),
-						};
-					});
-				return { nodes, error: null };
-			}
-		)
-	);
-
-	// Switch to an existing branch in the current working tree.
-	// Returns success=false with stderr text on failure (e.g., dirty working tree).
-	ipcMain.handle(
-		'git:switch',
-		withIpcErrorLogging(
-			handlerOpts('switch'),
-			async (cwd: string, branchName: string, sshRemoteId?: string, remoteCwd?: string) => {
-				// Reject flag-like names so a caller can't pass e.g. "-c new-branch" or "-C"
-				// and have git interpret it as a switch flag. execFile blocks shell
-				// injection but not flag injection.
-				if (
-					typeof branchName !== 'string' ||
-					branchName.length === 0 ||
-					branchName.startsWith('-')
-				) {
-					return {
-						success: false,
-						stdout: '',
-						stderr: `Invalid branch name: ${branchName}`,
-					};
-				}
-				const sshRemote = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-				const effectiveRemoteCwd = sshRemote ? remoteCwd || cwd : undefined;
-				const result = await execGit(['switch', branchName], cwd, sshRemote, effectiveRemoteCwd);
-				return {
-					success: result.exitCode === 0,
-					stdout: result.stdout,
-					stderr: result.stderr,
-				};
-			}
-		)
-	);
-
-	// --- Streaming remote-sync commands (pull / push / fetch) ---
-	//
-	// These are the only git operations the user watches happen: they hit the
-	// network, can take a while, and their progress output is the point. Instead
-	// of buffering, the child's stdout/stderr are forwarded chunk-by-chunk to the
-	// requesting window on `git:commandOutput` so the Git command modal can
-	// render a live console.
-	ipcMain.handle(
-		'git:runCommand',
-		async (
-			event,
-			options: {
-				runId: string;
-				operation: GitStreamingOperation;
-				cwd: string;
-				sshRemoteId?: string;
-				remoteCwd?: string;
-				/** push only: also set the upstream to origin/<current branch>. */
-				setUpstream?: boolean;
-			}
-		): Promise<GitRunCommandResult> => {
-			const { runId, operation, cwd, sshRemoteId, remoteCwd, setUpstream } = options;
-			try {
-				return await runStreamingGitCommand(event, {
-					runId,
-					operation,
-					cwd,
-					sshRemoteId,
-					remoteCwd,
-					setUpstream,
-				});
-			} catch (error) {
-				logger.error('runCommand error', LOG_CONTEXT, {
-					operation,
-					message: error instanceof Error ? error.message : String(error),
-				});
-				return {
-					success: false,
-					exitCode: 1,
-					cancelled: false,
-					error: error instanceof Error ? error.message : String(error),
-				};
-			}
-		}
-	);
-
-	// Terminate an in-flight `git:runCommand` (the modal's Cancel button).
-	ipcMain.handle(
-		'git:cancelCommand',
-		withIpcErrorLogging(handlerOpts('cancelCommand'), async (runId: string) => {
-			const cancel = streamingGitRuns.get(runId);
-			if (!cancel) return { success: false };
-			cancel();
-			return { success: true };
-		})
-	);
-
-	// Switch the working tree to another branch. `createTracking` checks out a
-	// remote-only branch (`git checkout -b <name> --track origin/<name>`), which
-	// is what the branch switcher needs for branches that exist only on origin.
-	ipcMain.handle(
-		'git:checkoutBranch',
-		withIpcErrorLogging(
-			handlerOpts('checkoutBranch'),
-			async (
-				cwd: string,
-				branch: string,
-				createTracking?: boolean,
-				sshRemoteId?: string,
-				remoteCwd?: string
-			) => {
-				const sshRemote = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-				// The user opted into SSH; silently checking out the local repo
-				// instead would switch the wrong working tree.
-				if (sshRemoteId && !sshRemote) {
-					return { success: false, error: `SSH remote not found: ${sshRemoteId}` };
-				}
-				const effectiveRemoteCwd = sshRemote ? remoteCwd || cwd : undefined;
-				const args = createTracking
-					? ['checkout', '-b', branch, '--track', `origin/${branch}`]
-					: ['checkout', branch];
-				const result = await execGit(args, cwd, sshRemote, effectiveRemoteCwd);
-				if (result.exitCode !== 0) {
-					return {
-						success: false,
-						error: result.stderr?.trim() || result.stdout?.trim() || 'git checkout failed',
-					};
-				}
-				// git checkout reports "Switched to branch 'x'" on stderr.
-				return { success: true, output: (result.stderr || result.stdout).trim() };
-			}
-		)
-	);
-
-	ipcMain.handle(
-		'git:commitCount',
-		withIpcErrorLogging(
-			handlerOpts('commitCount'),
-			async (cwd: string, sshRemoteId?: string, remoteCwd?: string) => {
-				const sshRemote = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-				const effectiveRemoteCwd = sshRemote ? remoteCwd || cwd : undefined;
-				// Get total commit count using rev-list
-				const result = await execGit(
-					['rev-list', '--count', 'HEAD'],
-					cwd,
-					sshRemote,
-					effectiveRemoteCwd
-				);
-				if (result.exitCode !== 0) {
-					return { count: 0, error: result.stderr };
-				}
-				return { count: parseInt(result.stdout.trim(), 10) || 0, error: null };
-			}
-		)
-	);
-
-	ipcMain.handle(
-		'git:show',
-		withIpcErrorLogging(
-			handlerOpts('show'),
-			async (cwd: string, hash: string, sshRemoteId?: string, remoteCwd?: string) => {
-				const sshRemote = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-				const effectiveRemoteCwd = sshRemote ? remoteCwd || cwd : undefined;
-				// Get the full diff for a specific commit
-				const result = await execGit(
-					['show', '--stat', '--patch', hash],
-					cwd,
-					sshRemote,
-					effectiveRemoteCwd
-				);
-				return { stdout: result.stdout, stderr: result.stderr };
-			}
-		)
-	);
-
-	// Read file content at a specific git ref (e.g., HEAD:path/to/file.png)
-	// Returns base64 data URL for images, raw content for text files
-	ipcMain.handle(
-		'git:showFile',
-		withIpcErrorLogging(
-			handlerOpts('showFile'),
-			async (cwd: string, ref: string, filePath: string) => {
-				// Use git show to get file content at specific ref
-				// We need to handle binary files differently
-				const ext = filePath.split('.').pop()?.toLowerCase() || '';
-
-				if (isImageFile(filePath)) {
-					// For images we need raw binary content. Use the async,
-					// binary-safe exec helper instead of spawnSync so reading the blob
-					// never blocks the main-process event loop (which would freeze the
-					// entire UI while a large image / cold git object is fetched).
-					const result = await execFileBufferNoThrow(
-						'git',
-						['show', `${ref}:${filePath}`],
-						cwd,
-						50 * 1024 * 1024 // 50MB max
-					);
-
-					if (result.exitCode !== 0) {
-						return { error: result.stderr || 'Failed to read file from git' };
-					}
-
-					const base64 = result.stdout.toString('base64');
-					const mimeType = getImageMimeType(ext);
-					return { content: `data:${mimeType};base64,${base64}` };
-				} else {
-					// For text files, use regular exec
-					const result = await execFileNoThrow('git', ['show', `${ref}:${filePath}`], cwd);
-					if (result.exitCode !== 0) {
-						return { error: result.stderr || 'Failed to read file from git' };
-					}
-					return { content: result.stdout };
-				}
-			}
-		)
-	);
-
-	// Git worktree operations for Auto Run parallelization
-
-	// Get information about a worktree at a given path
-	// Supports SSH remote execution via optional sshRemoteId parameter
-	ipcMain.handle(
-		'git:worktreeInfo',
-		createIpcHandler(
-			handlerOpts('worktreeInfo'),
-			async (worktreePath: string, sshRemoteId?: string) => {
-				// SSH remote: dispatch to remote git operations
-				if (sshRemoteId) {
-					const sshConfig = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-					if (!sshConfig) {
-						throw new Error(`SSH remote not found: ${sshRemoteId}`);
-					}
-					logger.debug(`${LOG_CONTEXT} worktreeInfo via SSH: ${worktreePath}`, LOG_CONTEXT);
-					const result = await worktreeInfoRemote(worktreePath, sshConfig);
-					if (!result.success || !result.data) {
-						throw new Error(result.error || 'Remote worktreeInfo failed');
-					}
-					return result.data;
-				}
-
-				// Local execution (existing code)
-				// Check if the path exists
-				try {
-					await fs.access(worktreePath);
-				} catch {
-					return { exists: false, isWorktree: false };
-				}
-
-				// Check if it's a git directory (could be main repo or worktree)
-				const isInsideWorkTree = await execFileNoThrow(
-					'git',
-					['rev-parse', '--is-inside-work-tree'],
-					worktreePath
-				);
-				if (isInsideWorkTree.exitCode !== 0) {
-					return { exists: true, isWorktree: false };
-				}
-
-				// Run git queries in parallel to reduce latency
-				const [gitDirResult, gitCommonDirResult, branchResult, repoRootResult] = await Promise.all([
-					execFileNoThrow('git', ['rev-parse', '--git-dir'], worktreePath),
-					execFileNoThrow('git', ['rev-parse', '--git-common-dir'], worktreePath),
-					execFileNoThrow('git', ['rev-parse', '--abbrev-ref', 'HEAD'], worktreePath),
-					execFileNoThrow('git', ['rev-parse', '--show-toplevel'], worktreePath),
-				]);
-				if (gitDirResult.exitCode !== 0) {
-					throw new Error('Failed to get git directory');
-				}
-				const gitDir = gitDirResult.stdout.trim();
-
-				const gitCommonDir =
-					gitCommonDirResult.exitCode === 0 ? gitCommonDirResult.stdout.trim() : gitDir;
-
-				// If git-dir and git-common-dir are different, this is a worktree
-				const isWorktree = gitDir !== gitCommonDir;
-
-				const currentBranch = branchResult.exitCode === 0 ? branchResult.stdout.trim() : undefined;
-
-				let repoRoot: string | undefined;
-
-				if (isWorktree && gitCommonDir) {
-					// For worktrees, we need to find the main repo root from the common dir
-					// The common dir points to the .git folder of the main repo
-					// The main repo root is the parent of the .git folder
-					const commonDirAbs = path.isAbsolute(gitCommonDir)
-						? gitCommonDir
-						: path.resolve(worktreePath, gitCommonDir);
-					repoRoot = path.dirname(commonDirAbs);
-				} else if (repoRootResult.exitCode === 0) {
-					repoRoot = repoRootResult.stdout.trim();
-				}
-
-				return {
-					exists: true,
-					isWorktree,
-					currentBranch,
-					repoRoot,
-				};
-			}
-		)
-	);
-
-	// Get the root directory of the git repository
-	// Supports SSH remote execution via optional sshRemoteId parameter
-	ipcMain.handle(
-		'git:getRepoRoot',
-		createIpcHandler(handlerOpts('getRepoRoot'), async (cwd: string, sshRemoteId?: string) => {
-			// SSH remote: dispatch to remote git operations
-			if (sshRemoteId) {
-				const sshConfig = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-				if (!sshConfig) {
-					throw new Error(`SSH remote not found: ${sshRemoteId}`);
-				}
-				logger.debug(`${LOG_CONTEXT} getRepoRoot via SSH: ${cwd}`, LOG_CONTEXT);
-				const result = await getRepoRootRemote(cwd, sshConfig);
-				if (!result.success) {
-					throw new Error(result.error || 'Not a git repository');
-				}
-				return { root: result.data };
-			}
-
-			// Local execution
-			const result = await execFileNoThrow('git', ['rev-parse', '--show-toplevel'], cwd);
-			if (result.exitCode !== 0) {
-				throw new Error(result.stderr || 'Not a git repository');
-			}
-			return { root: result.stdout.trim() };
-		})
-	);
-
-	// Create or reuse a worktree
-	// Supports SSH remote execution via optional sshRemoteId parameter
-	ipcMain.handle(
-		'git:worktreeSetup',
-		withIpcErrorLogging(
-			handlerOpts('worktreeSetup'),
-			async (
-				mainRepoCwd: string,
-				worktreePath: string,
-				branchName: string,
-				sshRemoteId?: string,
-				baseBranch?: string
-			) => {
-				// SSH remote: dispatch to remote git operations
-				if (sshRemoteId) {
-					const sshConfig = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-					if (!sshConfig) {
-						throw new Error(`SSH remote not found: ${sshRemoteId}`);
-					}
-					logger.debug(
-						`${LOG_CONTEXT} worktreeSetup via SSH: ${JSON.stringify({ mainRepoCwd, worktreePath, branchName, baseBranch })}`,
-						LOG_CONTEXT
-					);
-					const result = await worktreeSetupRemote(
-						mainRepoCwd,
-						worktreePath,
-						branchName,
-						sshConfig,
-						baseBranch
-					);
-					if (!result.success) {
-						throw new Error(result.error || 'Remote worktreeSetup failed');
-					}
-					return result.data;
-				}
-
-				// Local execution (existing code)
-				logger.debug(
-					`worktreeSetup called with: ${JSON.stringify({ mainRepoCwd, worktreePath, branchName, baseBranch })}`,
-					LOG_CONTEXT
-				);
-
-				// Resolve paths to absolute for proper comparison
-				const resolvedMainRepo = path.resolve(mainRepoCwd);
-				const resolvedWorktree = path.resolve(worktreePath);
-				logger.debug(
-					`Resolved paths: ${JSON.stringify({ resolvedMainRepo, resolvedWorktree })}`,
-					LOG_CONTEXT
-				);
-
-				// Check if worktree path is inside the main repo (nested worktree)
-				// This can cause issues because git and Claude Code search upward for .git
-				// and may resolve to the parent repo instead of the worktree
-				if (resolvedWorktree.startsWith(resolvedMainRepo + path.sep)) {
-					return {
-						success: false,
-						error:
-							'Worktree path cannot be inside the main repository. Please use a sibling directory (e.g., ../my-worktree) instead.',
-					};
-				}
-
-				// First check if the worktree path already exists
-				let pathExists = true;
-				try {
-					await fs.access(resolvedWorktree);
-					logger.debug(`Path exists: ${resolvedWorktree}`, LOG_CONTEXT);
-				} catch {
-					pathExists = false;
-					logger.debug(`Path does not exist: ${resolvedWorktree}`, LOG_CONTEXT);
-				}
-
-				if (pathExists) {
-					// Check if it's already a worktree of this repo
-					const worktreeInfoResult = await execFileNoThrow(
-						'git',
-						['rev-parse', '--is-inside-work-tree'],
-						resolvedWorktree
-					);
-					logger.debug(
-						`is-inside-work-tree result: ${JSON.stringify(worktreeInfoResult)}`,
-						LOG_CONTEXT
-					);
-					if (worktreeInfoResult.exitCode !== 0) {
-						// Path exists but isn't a git repo - check if it's empty and can be removed
-						const dirContents = await fs.readdir(resolvedWorktree);
-						logger.debug(`Directory contents: ${JSON.stringify(dirContents)}`, LOG_CONTEXT);
-						if (dirContents.length === 0) {
-							// Empty directory - remove it so we can create the worktree
-							logger.debug(`Removing empty directory`, LOG_CONTEXT);
-							await fs.rmdir(resolvedWorktree);
-							pathExists = false;
-						} else {
-							logger.debug(`Directory not empty, returning error`, LOG_CONTEXT);
-							return {
-								success: false,
-								error: 'Path exists but is not a git worktree or repository (and is not empty)',
-							};
-						}
-					}
-				}
-
-				if (pathExists) {
-					// Get the common dir to check if it's the same repo (parallel)
-					const [gitCommonDirResult, mainGitDirResult] = await Promise.all([
-						execFileNoThrow('git', ['rev-parse', '--git-common-dir'], resolvedWorktree),
-						execFileNoThrow('git', ['rev-parse', '--git-dir'], resolvedMainRepo),
-					]);
-
-					if (gitCommonDirResult.exitCode === 0 && mainGitDirResult.exitCode === 0) {
-						const worktreeCommonDir = path.resolve(
-							resolvedWorktree,
-							gitCommonDirResult.stdout.trim()
-						);
-						const mainGitDir = path.resolve(resolvedMainRepo, mainGitDirResult.stdout.trim());
-
-						// Normalize paths for comparison
-						const normalizedWorktreeCommon = path.normalize(worktreeCommonDir);
-						const normalizedMainGit = path.normalize(mainGitDir);
-
-						if (normalizedWorktreeCommon !== normalizedMainGit) {
-							return { success: false, error: 'Worktree path belongs to a different repository' };
-						}
-					}
-
-					// Get current branch in the existing worktree
-					const currentBranchResult = await execFileNoThrow(
-						'git',
-						['rev-parse', '--abbrev-ref', 'HEAD'],
-						worktreePath
-					);
-					const currentBranch =
-						currentBranchResult.exitCode === 0 ? currentBranchResult.stdout.trim() : '';
-
-					return {
-						success: true,
-						created: false,
-						currentBranch,
-						requestedBranch: branchName,
-						branchMismatch: currentBranch !== branchName && branchName !== '',
-					};
-				}
-
-				// Worktree doesn't exist, create it
-				// First check if the branch exists
-				const branchExistsResult = await execFileNoThrow(
-					'git',
-					['rev-parse', '--verify', branchName],
-					mainRepoCwd
-				);
-				const branchExists = branchExistsResult.exitCode === 0;
-
-				let createResult;
-				if (branchExists) {
-					// Branch exists, just add worktree pointing to it. baseBranch is
-					// ignored here because the existing branch already has its own commit.
-					createResult = await execFileNoThrow(
-						'git',
-						['worktree', 'add', worktreePath, branchName],
-						mainRepoCwd
-					);
-				} else if (baseBranch) {
-					// Branch doesn't exist; create it from the requested base branch.
-					// `git worktree add -b <new> <path> <base>` is the explicit form.
-					createResult = await execFileNoThrow(
-						'git',
-						['worktree', 'add', '-b', branchName, worktreePath, baseBranch],
-						mainRepoCwd
-					);
-				} else {
-					// Branch doesn't exist and no base specified; defaults to current HEAD
-					// of the main repo (preserves pre-baseBranch behavior).
-					createResult = await execFileNoThrow(
-						'git',
-						['worktree', 'add', '-b', branchName, worktreePath],
-						mainRepoCwd
-					);
-				}
-
-				if (createResult.exitCode !== 0) {
-					// Recover from "already used / already checked out" - the branch is
-					// already registered with another worktree on disk. Resolve that path
-					// from `git worktree list --porcelain` so the caller can open it.
-					const errMsg = createResult.stderr || '';
-					if (isWorktreeAlreadyUsedError(errMsg)) {
-						const existingPath = await findLocalWorktreeForBranch(mainRepoCwd, branchName);
-						if (existingPath) {
-							return {
-								success: true,
-								created: false,
-								alreadyExisted: true,
-								existingPath,
-								currentBranch: branchName,
-								requestedBranch: branchName,
-								branchMismatch: false,
-							};
-						}
-					}
-					return { success: false, error: createResult.stderr || 'Failed to create worktree' };
-				}
-
-				return {
-					success: true,
-					created: true,
-					currentBranch: branchName,
-					requestedBranch: branchName,
-					branchMismatch: false,
-				};
-			}
-		)
-	);
-
-	// Checkout a branch in a worktree (with uncommitted changes check)
-	// Supports SSH remote execution via optional sshRemoteId parameter
-	ipcMain.handle(
-		'git:worktreeCheckout',
-		withIpcErrorLogging(
-			handlerOpts('worktreeCheckout'),
-			async (
-				worktreePath: string,
-				branchName: string,
-				createIfMissing: boolean,
-				sshRemoteId?: string
-			) => {
-				// SSH remote: dispatch to remote git operations
-				if (sshRemoteId) {
-					const sshConfig = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-					if (!sshConfig) {
-						throw new Error(`SSH remote not found: ${sshRemoteId}`);
-					}
-					logger.debug(
-						`${LOG_CONTEXT} worktreeCheckout via SSH: ${JSON.stringify({ worktreePath, branchName, createIfMissing })}`,
-						LOG_CONTEXT
-					);
-					const result = await worktreeCheckoutRemote(
-						worktreePath,
-						branchName,
-						createIfMissing,
-						sshConfig
-					);
-					if (!result.success) {
-						throw new Error(result.error || 'Remote worktreeCheckout failed');
-					}
-					return result.data;
-				}
-
-				// Local execution (existing code)
-				// Check for uncommitted changes
-				const statusResult = await execFileNoThrow('git', ['status', '--porcelain'], worktreePath);
-				if (statusResult.exitCode !== 0) {
-					return {
-						success: false,
-						hasUncommittedChanges: false,
-						error: 'Failed to check git status',
-					};
-				}
-
-				const uncommittedChanges = statusResult.stdout.trim().length > 0;
-				if (uncommittedChanges) {
-					return {
-						success: false,
-						hasUncommittedChanges: true,
-						error: 'Worktree has uncommitted changes. Please commit or stash them first.',
-					};
-				}
-
-				// Check if branch exists
-				const branchExistsResult = await execFileNoThrow(
-					'git',
-					['rev-parse', '--verify', branchName],
-					worktreePath
-				);
-				const branchExists = branchExistsResult.exitCode === 0;
-
-				let checkoutResult;
-				if (branchExists) {
-					checkoutResult = await execFileNoThrow('git', ['checkout', branchName], worktreePath);
-				} else if (createIfMissing) {
-					checkoutResult = await execFileNoThrow(
-						'git',
-						['checkout', '-b', branchName],
-						worktreePath
-					);
-				} else {
-					return {
-						success: false,
-						hasUncommittedChanges: false,
-						error: `Branch '${branchName}' does not exist`,
-					};
-				}
-
-				if (checkoutResult.exitCode !== 0) {
-					return {
-						success: false,
-						hasUncommittedChanges: false,
-						error: checkoutResult.stderr || 'Checkout failed',
-					};
-				}
-
-				return { success: true, hasUncommittedChanges: false };
-			}
-		)
-	);
-
-	// Create a PR from the worktree branch to a base branch
-	// ghPath parameter allows specifying custom path to gh binary
-	ipcMain.handle(
-		'git:createPR',
-		withIpcErrorLogging(
-			handlerOpts('createPR'),
-			async (
-				worktreePath: string,
-				baseBranch: string,
-				title: string,
-				body: string,
-				ghPath?: string
-			) => {
-				// Resolve gh CLI path (uses cached detection or custom path)
-				const ghCommand = await resolveGhPath(ghPath);
-				logger.debug(`Using gh CLI at: ${ghCommand}`, LOG_CONTEXT);
-
-				// Build env with the user's full shell PATH so git hooks
-				// (e.g. Husky pre-push running npm) can find Node/npm binaries
-				let shellEnv: NodeJS.ProcessEnv | undefined;
-				try {
-					const shellPath = await getShellPath();
-					if (shellPath) {
-						shellEnv = { ...process.env, PATH: shellPath };
-					}
-				} catch (error) {
-					captureMessage(
-						`git:createPR falling back to default PATH: ${error instanceof Error ? error.message : String(error)}`,
-						'warning'
-					);
-				}
-
-				// First, push the current branch to origin
-				const pushResult = await execFileNoThrow(
-					'git',
-					['push', '-u', 'origin', 'HEAD'],
-					worktreePath,
-					shellEnv
-				);
-				if (pushResult.exitCode !== 0) {
-					return { success: false, error: `Failed to push branch: ${pushResult.stderr}` };
-				}
-
-				// Create the PR using gh CLI
-				const prResult = await execFileNoThrow(
-					ghCommand,
-					['pr', 'create', '--base', baseBranch, '--title', title, '--body', body],
-					worktreePath,
-					shellEnv
-				);
-
-				if (prResult.exitCode !== 0) {
-					// Check if gh CLI is not installed
-					if (
-						prResult.stderr.includes('command not found') ||
-						prResult.stderr.includes('not recognized')
-					) {
-						return {
-							success: false,
-							error: 'GitHub CLI (gh) is not installed. Please install it to create PRs.',
-						};
-					}
-					return { success: false, error: prResult.stderr || 'Failed to create PR' };
-				}
-
-				// The PR URL is typically in stdout
-				const prUrl = prResult.stdout.trim();
-				return { success: true, prUrl };
-			}
-		)
-	);
-
-	// Check if GitHub CLI (gh) is installed and authenticated
-	// ghPath parameter allows specifying custom path to gh binary (e.g., /opt/homebrew/bin/gh)
-	// Results are cached for 1 minute to avoid repeated subprocess calls
-	ipcMain.handle(
-		'git:checkGhCli',
-		withIpcErrorLogging(handlerOpts('checkGhCli'), async (ghPath?: string) => {
-			// Check cache first (skip if custom path provided)
-			if (!ghPath) {
-				const cached = getCachedGhStatus();
-				if (cached !== null) {
-					logger.debug(
-						`Using cached gh CLI status: installed=${cached.installed}, authenticated=${cached.authenticated}`,
-						LOG_CONTEXT
-					);
-					return cached;
-				}
-			}
-
-			// Resolve gh CLI path (uses cached detection or custom path)
-			const ghCommand = await resolveGhPath(ghPath);
-			logger.debug(`Checking gh CLI at: ${ghCommand}`, LOG_CONTEXT);
-
-			// Check if gh is installed by running gh --version
-			const versionResult = await execFileNoThrow(ghCommand, ['--version']);
-			if (versionResult.exitCode !== 0) {
-				logger.warn(
-					`gh CLI not found at ${ghCommand}: exit=${versionResult.exitCode}, stderr=${versionResult.stderr}`,
-					LOG_CONTEXT
-				);
-				const result = { installed: false, authenticated: false };
-				if (!ghPath) setCachedGhStatus(false, false);
-				return result;
-			}
-			logger.debug(`gh CLI found: ${versionResult.stdout.trim().split('\n')[0]}`, LOG_CONTEXT);
-
-			// Check if gh is authenticated by running gh auth status
-			const authResult = await execFileNoThrow(ghCommand, ['auth', 'status']);
-			const authenticated = authResult.exitCode === 0;
-			logger.debug(
-				`gh auth status: ${authenticated ? 'authenticated' : 'not authenticated'}`,
-				LOG_CONTEXT
-			);
-
-			// Cache the result (only if not using custom path)
-			if (!ghPath) {
-				setCachedGhStatus(true, authenticated);
-			}
-
-			return { installed: true, authenticated };
-		})
-	);
-
-	// Get the default branch name (main or master)
-	ipcMain.handle(
-		'git:getDefaultBranch',
-		createIpcHandler(handlerOpts('getDefaultBranch'), async (cwd: string) => {
-			// First try to get the default branch from remote
-			const remoteResult = await execFileNoThrow('git', ['remote', 'show', 'origin'], cwd);
-			if (remoteResult.exitCode === 0) {
-				// Parse "HEAD branch: main" from the output
-				const match = remoteResult.stdout.match(/HEAD branch:\s*(\S+)/);
-				if (match) {
-					return { branch: match[1] };
-				}
-			}
-
-			// Fallback: check if main or master exists locally
-			const mainResult = await execFileNoThrow('git', ['rev-parse', '--verify', 'main'], cwd);
-			if (mainResult.exitCode === 0) {
-				return { branch: 'main' };
-			}
-
-			const masterResult = await execFileNoThrow('git', ['rev-parse', '--verify', 'master'], cwd);
-			if (masterResult.exitCode === 0) {
-				return { branch: 'master' };
-			}
-
-			throw new Error('Could not determine default branch');
-		})
-	);
-
-	// List all worktrees for a git repository
-	// Supports SSH remote execution via optional sshRemoteId parameter
-	ipcMain.handle(
-		'git:listWorktrees',
-		createIpcHandler(handlerOpts('listWorktrees'), async (cwd: string, sshRemoteId?: string) => {
-			// SSH remote: dispatch to remote git operations
-			if (sshRemoteId) {
-				const sshConfig = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-				if (!sshConfig) {
-					throw new Error(`SSH remote not found: ${sshRemoteId}`);
-				}
-				logger.debug(`${LOG_CONTEXT} listWorktrees via SSH: ${cwd}`, LOG_CONTEXT);
-				const result = await listWorktreesRemote(cwd, sshConfig);
-				if (!result.success) {
-					throw new Error(result.error || 'Remote listWorktrees failed');
-				}
-				return { worktrees: result.data };
-			}
-
-			// Local execution (existing code)
-			// Run git worktree list --porcelain for machine-readable output
-			const result = await execFileNoThrow('git', ['worktree', 'list', '--porcelain'], cwd);
-			if (result.exitCode !== 0) {
-				// Not a git repo or no worktree support
-				return { worktrees: [] };
-			}
-
-			// Parse porcelain output:
-			// worktree /path/to/worktree
-			// HEAD abc123
-			// branch refs/heads/branch-name
-			// (blank line separates entries)
-			const worktrees: Array<{
-				path: string;
-				head: string;
-				branch: string | null;
-				isBare: boolean;
-			}> = [];
-
-			const lines = result.stdout.split('\n');
-			let current: { path?: string; head?: string; branch?: string | null; isBare?: boolean } = {};
-
-			for (const line of lines) {
-				if (line.startsWith('worktree ')) {
-					current.path = line.substring(9);
-				} else if (line.startsWith('HEAD ')) {
-					current.head = line.substring(5);
-				} else if (line.startsWith('branch ')) {
-					// Extract branch name from refs/heads/branch-name
-					const branchRef = line.substring(7);
-					current.branch = branchRef.replace('refs/heads/', '');
-				} else if (line === 'bare') {
-					current.isBare = true;
-				} else if (line === 'detached') {
-					current.branch = null; // Detached HEAD
-				} else if (line === '' && current.path) {
-					// End of entry
-					worktrees.push({
-						path: current.path,
-						head: current.head || '',
-						branch: current.branch ?? null,
-						isBare: current.isBare || false,
-					});
-					current = {};
-				}
-			}
-
-			// Handle last entry if no trailing newline
-			if (current.path) {
-				worktrees.push({
-					path: current.path,
-					head: current.head || '',
-					branch: current.branch ?? null,
-					isBare: current.isBare || false,
-				});
-			}
-
-			return { worktrees };
-		})
-	);
-
-	// Scan a directory for subdirectories that are git repositories or worktrees
-	// This is used for auto-discovering worktrees in a parent directory
-	// PERFORMANCE: Parallelized git operations to avoid blocking UI (was sequential before)
-	// Supports SSH remote execution via optional sshRemoteId parameter
-	//
-	// Recurses one level into non-git subdirectories so worktrees created from
-	// branch names with slashes (e.g. "fix/worktree-removal" → /worktrees/fix/worktree-removal)
-	// are still discovered. Without recursion, those nested worktrees are absent
-	// from the result and the renderer's stale-detection wrongly removes them.
-	ipcMain.handle(
-		'git:scanWorktreeDirectory',
-		createIpcHandler(
-			handlerOpts('scanWorktreeDirectory'),
-			async (parentPath: string, sshRemoteId?: string) => {
-				const sshRemote = sshRemoteId ? getSshRemoteById(sshRemoteId) : undefined;
-
-				// Maximum recursion depth below the configured basePath. 1 covers the
-				// common case `<basePath>/<group>/<branch>` from slash-named branches.
-				// Going deeper would multiply git invocations without a real-world need
-				// (git itself rejects nested worktrees inside the main repo).
-				const MAX_DEPTH = 1;
-
-				type SubdirEntry = { name: string; isDirectory: boolean };
-				type ScanEntry = {
-					path: string;
-					name: string;
-					isWorktree: boolean;
-					branch: string | null;
-					repoRoot: string | null;
-				};
-
-				const joinPath = (parent: string, child: string): string =>
-					sshRemote
-						? parent.endsWith('/')
-							? `${parent}${child}`
-							: `${parent}/${child}`
-						: path.join(parent, child);
-
-				// Throws on read failure (matching local `fs.readdir` behavior) so the
-				// outer try/catch can surface scanFailed: true at the top level. Nested
-				// recursion wraps this in its own try/catch and swallows the throw.
-				// Without this, an SSH `readDirRemote` failure would silently return []
-				// and the renderer would bulk-remove every child session.
-				const readSubdirs = async (dir: string): Promise<SubdirEntry[]> => {
-					if (sshRemote) {
-						const result = await readDirRemote(dir, sshRemote);
-						if (!result.success || !result.data) {
-							const err = new Error(
-								`Failed to read remote directory ${dir}: ${result.error || 'unknown error'}`
-							) as NodeJS.ErrnoException;
-							// Tag as ENOENT so the outer catch's Sentry-quieting branch applies -
-							// remote read failures are typically "path no longer exists / not reachable",
-							// not bugs worth paging on.
-							err.code = 'ENOENT';
-							throw err;
-						}
-						return result.data.filter((e) => e.isDirectory && !e.name.startsWith('.'));
-					}
-					const entries = await fs.readdir(dir, { withFileTypes: true });
-					return entries
-						.filter((e) => e.isDirectory() && !e.name.startsWith('.'))
-						.map((e) => ({ name: e.name, isDirectory: true }));
-				};
-
-				// Inspect a single directory: returns the worktree entry if it IS a
-				// git repo/worktree root, or null otherwise. Caller decides whether
-				// to recurse into a null result.
-				const inspectSubdir = async (
-					subdirPath: string,
-					name: string
-				): Promise<ScanEntry | null> => {
-					const isInsideWorkTree = await execGit(
-						['rev-parse', '--is-inside-work-tree'],
-						subdirPath,
-						sshRemote
-					);
-					if (isInsideWorkTree.exitCode !== 0) {
-						return null; // Not a git repo
-					}
-
-					// Verify this directory IS a worktree/repo root, not just a subdirectory inside one.
-					// Without this check, subdirectories like "build/" or "src/" inside a worktree
-					// would pass --is-inside-work-tree and be incorrectly treated as separate worktrees.
-					const toplevelResult = await execGit(
-						['rev-parse', '--show-toplevel'],
-						subdirPath,
-						sshRemote
-					);
-					if (toplevelResult.exitCode !== 0) {
-						return null; // Git command failed - treat as invalid
-					}
-					const toplevel = toplevelResult.stdout.trim();
-					// For local paths, canonicalize via realpath so that symlinked base
-					// paths (common on Linux: /home → /data/home; Windows junctions) match
-					// what git rev-parse --show-toplevel returns. path.resolve alone does
-					// NOT follow symlinks, which previously caused every subdir to be
-					// rejected and the entire worktree set to be marked stale.
-					const normalizedSubdir = sshRemote
-						? subdirPath
-						: await fs.realpath(subdirPath).catch(() => path.resolve(subdirPath));
-					const normalizedToplevel = sshRemote
-						? toplevel
-						: await fs.realpath(toplevel).catch(() => path.resolve(toplevel));
-					if (normalizedSubdir !== normalizedToplevel) {
-						return null; // Subdirectory inside a repo, not a repo/worktree root
-					}
-
-					// Run remaining git commands in parallel for each subdirectory (SSH-aware via execGit)
-					const [gitDirResult, gitCommonDirResult, branchResult] = await Promise.all([
-						execGit(['rev-parse', '--git-dir'], subdirPath, sshRemote),
-						execGit(['rev-parse', '--git-common-dir'], subdirPath, sshRemote),
-						execGit(['rev-parse', '--abbrev-ref', 'HEAD'], subdirPath, sshRemote),
-					]);
-
-					const gitDir = gitDirResult.exitCode === 0 ? gitDirResult.stdout.trim() : '';
-					const gitCommonDir =
-						gitCommonDirResult.exitCode === 0 ? gitCommonDirResult.stdout.trim() : gitDir;
-					const isWorktree = gitDir !== gitCommonDir;
-					const branch = branchResult.exitCode === 0 ? branchResult.stdout.trim() : null;
-
-					// Get repo root
-					let repoRoot: string | null = null;
-					if (isWorktree && gitCommonDir) {
-						// For SSH, use POSIX path operations
-						if (sshRemote) {
-							const commonDirAbs = gitCommonDir.startsWith('/')
-								? gitCommonDir
-								: `${subdirPath}/${gitCommonDir}`.replace(/\/+/g, '/');
-							// Get parent directory (remove last path component)
-							repoRoot = commonDirAbs.split('/').slice(0, -1).join('/') || '/';
-						} else {
-							const commonDirAbs = path.isAbsolute(gitCommonDir)
-								? gitCommonDir
-								: path.resolve(subdirPath, gitCommonDir);
-							repoRoot = path.dirname(commonDirAbs);
-						}
-					} else {
-						// For non-worktree git repos, the toplevel IS the repo root -
-						// reuse the value we already fetched above instead of re-running
-						// `git rev-parse --show-toplevel`.
-						repoRoot = toplevel;
-					}
-
-					return {
-						path: subdirPath,
-						name,
-						isWorktree,
-						branch,
-						repoRoot,
-					};
-				};
-
-				// Walk a directory level: inspect each subdir, then recurse into any
-				// non-git subdirs (up to MAX_DEPTH below the original parentPath).
-				// Failures while reading a nested directory are swallowed by the
-				// inner try/catch - a missing or unreadable group dir shouldn't fail
-				// the entire scan. Top-level failure propagates up to the outer
-				// try/catch so scanFailed is surfaced and the renderer skips removal.
-				const scanLevel = async (dir: string, depthRemaining: number): Promise<ScanEntry[]> => {
-					const subdirs = await readSubdirs(dir);
-
-					const results = await Promise.all(
-						subdirs.map(async (subdir) => {
-							const subdirPath = joinPath(dir, subdir.name);
-							const entry = await inspectSubdir(subdirPath, subdir.name);
-							if (entry) {
-								return [entry];
-							}
-							if (depthRemaining > 0) {
-								try {
-									return await scanLevel(subdirPath, depthRemaining - 1);
-								} catch (err) {
-									const code = (err as NodeJS.ErrnoException | undefined)?.code;
-									if (!isExpectedScanError(code)) {
-										logger.warn(`${LOG_CONTEXT} Failed to recurse into ${subdirPath}: ${err}`);
-									}
-									return [];
-								}
-							}
-							return [];
-						})
-					);
-
-					return results.flat();
-				};
-
-				try {
-					const gitSubdirs = await scanLevel(parentPath, MAX_DEPTH);
-					return { gitSubdirs };
-				} catch (err) {
-					// The configured parent path is user-supplied, so failing to read it is
-					// an environment condition rather than a bug: ENOENT when it's been moved
-					// or deleted, and EPERM/EACCES when it sits behind macOS TCC (Documents,
-					// Desktop) or has permissions we simply don't hold. Both surface to logs
-					// and to `scanFailed` below; neither should pollute Sentry. (MAESTRO-VQ)
-					const code = (err as NodeJS.ErrnoException | undefined)?.code;
-					if (!isExpectedScanError(code)) {
-						void captureException(err);
-					}
-					logger.error(`Failed to scan directory ${parentPath}: ${err}`, LOG_CONTEXT);
-					// Distinguish a failed scan from a successful "no subdirs" result so
-					// the renderer doesn't bulk-flag every existing child session as removed.
-					return { gitSubdirs: [], scanFailed: true };
-				}
-			}
-		)
-	);
-
-	// Watch a worktree directory for new worktrees
-	// Note: File watching is not supported for SSH remote sessions.
-	// Remote sessions will get success: true but isRemote: true flag indicating
-	// watching is not active. The UI should periodically poll listWorktrees instead.
-	ipcMain.handle(
-		'git:watchWorktreeDirectory',
-		createIpcHandler(
-			handlerOpts('watchWorktreeDirectory'),
-			async (sessionId: string, worktreePath: string, sshRemoteId?: string) => {
-				// TODO: Remove debug logging after worktree detection is confirmed working
-				logger.warn(
-					`[WT-DEBUG] watchWorktreeDirectory called: session=${sessionId} path=${worktreePath} ssh=${sshRemoteId}`
-				);
-
-				// SSH remote: file watching is not supported
-				// Return success with isRemote flag so UI knows to poll instead
-				if (sshRemoteId) {
-					logger.debug(
-						`${LOG_CONTEXT} Worktree watching not supported for SSH remote sessions. Session ${sessionId} should poll instead.`,
-						LOG_CONTEXT
-					);
-					return {
-						success: true,
-						isRemote: true,
-						message: 'File watching not available for remote sessions. Use polling instead.',
-					};
-				}
-
-				// Stop existing watcher if any - delete from map BEFORE awaiting close
-				// to prevent race conditions with concurrent unwatch/watch IPC calls
-				const existingWatcher = worktreeWatchers.get(sessionId);
-				if (existingWatcher) {
-					worktreeWatchers.delete(sessionId);
-					await existingWatcher.close();
-				}
-
-				// Clear any pending debounce timers for this session
-				for (const [key, timer] of worktreeWatchDebounceTimers) {
-					if (key.startsWith(`${sessionId}:`)) {
-						clearTimeout(timer);
-						worktreeWatchDebounceTimers.delete(key);
-					}
-				}
-
-				try {
-					// Verify directory exists
-					await fs.access(worktreePath);
-
-					// Watch one level deep so worktrees from slash-named branches
-					// (e.g. "fix/worktree-removal" → <basePath>/fix/worktree-removal)
-					// also fire addDir/unlinkDir events. The addDir handler validates
-					// every candidate via `is-inside-work-tree` + `show-toplevel`, so
-					// the intermediate group directory (e.g. "fix") is rejected and
-					// only the actual worktree is reported as discovered.
-					const watcher = chokidar.watch(worktreePath, {
-						ignored: [
-							/(^|[/\\])\../, // Ignore dotfiles
-							WINDOWS_LOCKED_SYSTEM_FILES,
-						],
-						persistent: true,
-						ignoreInitial: true,
-						depth: 1,
-					});
-
-					// Handler for directory additions
-					watcher.on('addDir', async (dirPath: string) => {
-						// TODO: Remove debug logging after worktree detection is confirmed working
-						logger.warn(`[WT-DEBUG] addDir event: ${dirPath}`);
-						// Skip the root directory itself
-						if (dirPath === worktreePath) return;
-
-						// Per-directory debounce so multiple near-simultaneous worktree
-						// additions each get their own validation pipeline
-						const debounceKey = `${sessionId}:${dirPath}`;
-						const existingTimer = worktreeWatchDebounceTimers.get(debounceKey);
-						if (existingTimer) {
-							clearTimeout(existingTimer);
-						}
-
-						const timer = setTimeout(async () => {
-							worktreeWatchDebounceTimers.delete(debounceKey);
-
-							// Check if this new directory is a git worktree
-							const isInsideWorkTree = await execFileNoThrow(
-								'git',
-								['rev-parse', '--is-inside-work-tree'],
-								dirPath
-							);
-							if (isInsideWorkTree.exitCode !== 0) {
-								logger.warn(
-									`[WT-DEBUG] REJECTED ${dirPath}: not inside work tree (exit=${isInsideWorkTree.exitCode} stderr=${isInsideWorkTree.stderr})`
-								);
-								return;
-							}
-
-							// Verify this IS a worktree/repo root, not a subdirectory inside one
-							const toplevelResult = await execFileNoThrow(
-								'git',
-								['rev-parse', '--show-toplevel'],
-								dirPath
-							);
-							if (toplevelResult.exitCode !== 0) {
-								logger.warn(
-									`[WT-DEBUG] REJECTED ${dirPath}: show-toplevel failed (exit=${toplevelResult.exitCode})`
-								);
-								return;
-							}
-							// Use realpath so symlinked base paths (e.g. /home/user/work →
-							// /data/work on Linux, NTFS junctions on Windows) match git's
-							// canonical toplevel output.
-							const resolvedDir = await fs.realpath(dirPath).catch(() => path.resolve(dirPath));
-							const resolvedToplevel = await fs
-								.realpath(toplevelResult.stdout.trim())
-								.catch(() => path.resolve(toplevelResult.stdout.trim()));
-							if (resolvedDir !== resolvedToplevel) {
-								logger.warn(
-									`[WT-DEBUG] REJECTED ${dirPath}: not repo root (resolved=${resolvedDir} toplevel=${resolvedToplevel})`
-								);
-								return;
-							}
-
-							// Get branch name
-							const branchResult = await execFileNoThrow(
-								'git',
-								['rev-parse', '--abbrev-ref', 'HEAD'],
-								dirPath
-							);
-							const branch = branchResult.exitCode === 0 ? branchResult.stdout.trim() : null;
-
-							// Skip main/master/HEAD branches
-							if (branch === 'main' || branch === 'master' || branch === 'HEAD') {
-								logger.warn(`[WT-DEBUG] REJECTED ${dirPath}: skippable branch ${branch}`);
-								return;
-							}
-
-							logger.warn(
-								`[WT-DEBUG] ACCEPTED ${dirPath}: branch=${branch}, emitting worktree:discovered`
-							);
-
-							// Emit event to the renderer and web-desktop bridge clients
-							safeSend('worktree:discovered', {
-								sessionId,
-								worktree: {
-									path: dirPath,
-									name: path.basename(dirPath),
-									branch,
-								},
-							});
-
-							logger.info(`${LOG_CONTEXT} New worktree discovered: ${dirPath} (branch: ${branch})`);
-						}, 500); // 500ms debounce
-
-						worktreeWatchDebounceTimers.set(debounceKey, timer);
-					});
-
-					// Handler for directory removals (e.g., `git worktree remove` from CLI).
-					//
-					// With depth: 1 this can fire spuriously for an intermediate group
-					// directory (e.g. <basePath>/fix) when its last nested worktree is
-					// removed and the empty parent is cleaned up. We forward the event
-					// regardless because (a) the dir is gone so we can't run git checks
-					// to validate, and (b) the renderer's onWorktreeRemoved handler
-					// already filters by registered child cwds - an unknown path is a
-					// no-op, not a session removal. See useWorktreeHandlers.ts.
-					watcher.on('unlinkDir', (dirPath: string) => {
-						if (dirPath === worktreePath) return;
-
-						logger.warn(`[WT-DEBUG] unlinkDir event: ${dirPath}`);
-						logger.info(`${LOG_CONTEXT} Worktree directory removed: ${dirPath}`);
-
-						// A deleted worktree is no longer a valid jump/diff target; flag
-						// any non-terminal agent runs bound to it as stale (ISC-6.7/D12).
-						markStaleForDeletedWorktreeUsingStore(dirPath);
-
-						safeSend('worktree:removed', {
-							sessionId,
-							worktreePath: dirPath,
-						});
-					});
-
-					watcher.on('error', (error) => {
-						logger.error(
-							`${LOG_CONTEXT} Worktree watcher error for session ${sessionId}: ${error}`
-						);
-					});
-
-					worktreeWatchers.set(sessionId, watcher);
-					logger.info(
-						`${LOG_CONTEXT} Started watching worktree directory: ${worktreePath} for session ${sessionId}`
-					);
-
-					return { success: true };
-				} catch (err) {
-					// ENOENT is expected when the worktree parent path has been moved
-					// or deleted; the renderer surfaces this as "stale" - no need to
-					// page Sentry on user filesystem state.
-					const code = (err as NodeJS.ErrnoException | undefined)?.code;
-					if (code !== 'ENOENT') {
-						void captureException(err);
-					}
-					logger.error(`${LOG_CONTEXT} Failed to watch worktree directory ${worktreePath}: ${err}`);
-					return { success: false, error: String(err) };
-				}
-			}
-		)
-	);
-
-	// Stop watching a worktree directory
-	ipcMain.handle(
-		'git:unwatchWorktreeDirectory',
-		createIpcHandler(handlerOpts('unwatchWorktreeDirectory'), async (sessionId: string) => {
-			// TODO: Remove debug logging after worktree detection is confirmed working
-			logger.warn(
-				`[WT-DEBUG] unwatchWorktreeDirectory called: session=${sessionId} hasWatcher=${worktreeWatchers.has(sessionId)}`
-			);
-			const watcher = worktreeWatchers.get(sessionId);
-			if (watcher) {
-				// Delete from map BEFORE awaiting close to prevent a race condition:
-				// React StrictMode double-fires effects, so unwatchWorktreeDirectory and
-				// watchWorktreeDirectory can interleave. If we delete after await, the
-				// unwatch can remove a NEW watcher that watchWorktreeDirectory just created.
-				worktreeWatchers.delete(sessionId);
-				await watcher.close();
-				logger.info(`${LOG_CONTEXT} Stopped watching worktree directory for session ${sessionId}`);
-			}
-
-			// Clear any pending debounce timers for this session
-			for (const [key, timer] of worktreeWatchDebounceTimers) {
-				if (key.startsWith(`${sessionId}:`)) {
-					clearTimeout(timer);
-					worktreeWatchDebounceTimers.delete(key);
-				}
-			}
-
-			return { success: true };
-		})
-	);
-
-	// Remove a worktree directory from disk
-	// Uses `git worktree remove` if it's a git worktree, or falls back to recursive delete
-	ipcMain.handle(
-		'git:removeWorktree',
-		withIpcErrorLogging(
-			handlerOpts('removeWorktree'),
-			async (worktreePath: string, force: boolean = false) => {
-				try {
-					// First check if the directory exists
-					await fs.access(worktreePath);
-
-					// Try to use git worktree remove first (cleanest approach)
-					const args = force
-						? ['worktree', 'remove', '--force', worktreePath]
-						: ['worktree', 'remove', worktreePath];
-					const gitResult = await execFileNoThrow('git', args, worktreePath);
-
-					if (gitResult.exitCode === 0) {
-						logger.info(`${LOG_CONTEXT} Removed worktree via git: ${worktreePath}`);
-						markStaleForDeletedWorktreeUsingStore(worktreePath);
-						return { success: true };
-					}
-
-					// If git worktree remove failed (maybe not a worktree or has changes), try force removal
-					if (!force) {
-						// Check if there are uncommitted changes
-						const statusResult = await execFileNoThrow(
-							'git',
-							['status', '--porcelain'],
-							worktreePath
-						);
-						if (statusResult.exitCode === 0 && statusResult.stdout.trim().length > 0) {
-							return {
-								success: false,
-								error: 'Worktree has uncommitted changes. Use force option to delete anyway.',
-								hasUncommittedChanges: true,
-							};
-						}
-					}
-
-					// Fall back to recursive directory removal
-					await fs.rm(worktreePath, { recursive: true, force: true });
-					logger.info(`${LOG_CONTEXT} Removed worktree directory: ${worktreePath}`);
-					markStaleForDeletedWorktreeUsingStore(worktreePath);
-					return { success: true };
-				} catch (err) {
-					const errorMessage = err instanceof Error ? err.message : String(err);
-					logger.error(`${LOG_CONTEXT} Failed to remove worktree ${worktreePath}: ${errorMessage}`);
-					return { success: false, error: errorMessage };
-				}
-			}
-		)
-	);
-
-	// Create a GitHub Gist from file content
-	// Returns the gist URL on success
-	ipcMain.handle(
-		'git:createGist',
-		withIpcErrorLogging(
-			handlerOpts('createGist'),
-			async (
-				filename: string,
-				content: string,
-				description: string,
-				isPublic: boolean,
-				ghPath?: string
-			) => {
-				// Resolve gh CLI path (uses cached detection or custom path)
-				const ghCommand = await resolveGhPath(ghPath);
-				logger.debug(`Using gh CLI for gist creation at: ${ghCommand}`, LOG_CONTEXT);
-
-				// Create gist using gh CLI with stdin for content
-				// gh gist create --filename <name> --desc <desc> [--public] -
-				const args = ['gist', 'create', '--filename', filename];
-				if (description) {
-					args.push('--desc', description);
-				}
-				if (isPublic) {
-					args.push('--public');
-				}
-				args.push('-'); // Read from stdin
-
-				const gistResult = await execFileNoThrow(ghCommand, args, undefined, { input: content });
-
-				if (gistResult.exitCode !== 0) {
-					// Check if gh CLI is not installed
-					if (
-						gistResult.stderr.includes('command not found') ||
-						gistResult.stderr.includes('not recognized')
-					) {
-						return {
-							success: false,
-							error: 'GitHub CLI (gh) is not installed. Please install it to create gists.',
-						};
-					}
-					// Check for authentication issues
-					if (
-						gistResult.stderr.includes('not logged') ||
-						gistResult.stderr.includes('authentication')
-					) {
-						return {
-							success: false,
-							error: 'GitHub CLI is not authenticated. Please run "gh auth login" first.',
-						};
-					}
-					return { success: false, error: gistResult.stderr || 'Failed to create gist' };
-				}
-
-				// The gist URL is typically in stdout
-				const gistUrl = gistResult.stdout.trim();
-				logger.info(`${LOG_CONTEXT} Created gist: ${gistUrl}`);
-				return { success: true, gistUrl };
-			}
-		)
-	);
-
-	logger.debug(`${LOG_CONTEXT} Git IPC handlers registered`);
-}
